backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
crewai>=0.102.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.25.2
requests>=2.31.0
python-multipart>=0.0.6
langchain-openai>=0.3.28
redis>=6.0.0
huey>=2.5.0